requirements.txt
pygame-ce
pygame_gui
transformers
torch
Pillow
ultralytics
